% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{acf_plot}
\alias{acf_plot}
\title{Plot the autocorrelation for each chain}
\usage{
acf_plot(fit, param)
}
\arguments{
\item{fit}{A footstanfit object.}

\item{param}{Desired param to be graphed out}
}
\value{
A GGplot plot
}
\description{
Plot the autocorrelation for each chain
}
