% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{dens_plot}
\alias{dens_plot}
\title{Plot the distribution of the draws for each chain}
\usage{
dens_plot(fit, param)
}
\arguments{
\item{fit}{A footstanfit object.}

\item{params}{Desired params to be graphed out}
}
\value{
A GGplot plot
}
\description{
Plot the distribution of the draws for each chain
}
