% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eliteserien}
\alias{eliteserien}
\title{Eliteserien, 2009 - 2018}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2335 rows and 9 columns.}
\source{
\href{https://www.eliteserien.no/resultater}{Eliteserien's home page}
}
\usage{
eliteserien
}
\description{
Data from the top flight of the Norwegian football league system. Contains
    goals scored for each game from the 2009 season to the middle of the 2018
    season.
}
\examples{
eliteserien

}
\keyword{datasets}
