% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ergotic_plot}
\alias{ergotic_plot}
\title{Plot ergotic mean of a param}
\usage{
ergotic_plot(fit, param)
}
\arguments{
\item{fit}{A footstanfit object.}

\item{params}{Desired params to be graphed out}
}
\value{
A GGplot plot
}
\description{
Plot ergotic mean of a param
}
