% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_data.R
\name{fit_data}
\alias{fit_data}
\title{Bayesian regression for football championships data}
\usage{
fit_data(data, hyperparams, model, ...)
}
\arguments{
\item{data}{List with the championship data.}

\item{hyperparams}{List of hyper parameters for the desired model.}

\item{model}{Model of choice.}

\item{...}{Arguments passed to \code{rstan::sampling} (e.g. iter, chains).}
}
\value{
An object of class \code{stanfit} returned by \code{rstan::sampling}
}
\description{
Bayesian regression for football championships data
}
