% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_data.R
\name{fit_poisson}
\alias{fit_poisson}
\title{Bayesian regression for football championships data using the poisson model}
\usage{
fit_poisson(
  data,
  beta_0_hyp,
  home_hyp,
  att_sd_hyp,
  def_sd_hyp,
  iter,
  chains,
  ...
)
}
\arguments{
\item{data}{Football games data}

\item{beta_0_hyp}{Vector of hyperparameters for beta_0 ~ N(beta_0_hyp\verb{[1]}, beta_0_hyp\verb{[2]})}

\item{home_hyp}{Vector of hyperparameters for home ~ N(home_hyp\verb{[1]}, home_hyp\verb{[2]})}

\item{att_sd_hyp}{Vector of hyperparameters for att_sd ~ Cauchy(att_sd_hyp\verb{[1]}, att_sd_hyp\verb{[2]})}

\item{def_sd_hyp}{Vector of hyperparameters for def_sd ~ Cauchy(def_sd_hyp\verb{[1]}, def_sd_hyp\verb{[2]})}

\item{iter}{Number of iterations used by \code{rstan::sampling}}

\item{...}{Other arguments passed to \code{rstan::sampling}}

\item{chain}{Number of chains used by \code{rstan::sampling}}
}
\value{
An object of class \code{stanfit} returned by \code{rstan::sampling}
}
\description{
Bayesian regression for football championships data using the poisson model
}
