% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_data.R
\name{fit_poisson_infl_2}
\alias{fit_poisson_infl_2}
\title{Bayesian regression for football championships data using a second negative binomial model}
\usage{
fit_poisson_infl_2(
  data,
  beta_0_hyp,
  home_hyp,
  att_h_sd_hyp,
  att_a_sd_hyp,
  def_h_sd_hyp,
  def_a_sd_hyp,
  p_zero_home_hyp,
  p_zero_away_hyp,
  iter,
  chains,
  ...
)
}
\arguments{
\item{data}{Football games data}

\item{beta_0_hyp}{Vector of hyperparameters for beta_0 ~ N(beta_0_hyp\verb{[1]}, beta_0_hyp\verb{[2]})}

\item{home_hyp}{Vector of hyperparameters for home ~ N(home_hyp\verb{[1]}, home_hyp\verb{[2]})}

\item{att_h_sd_hyp}{Vector of hyperparameters for att_h_sd ~ Cauchy(att_h_sd_hyp\verb{[1]}, att_h_sd_hyp\verb{[2]})}

\item{att_a_sd_hyp}{Vector of hyperparameters for att_a_sd ~ Cauchy(att_a_sd_hyp\verb{[1]}, att_a_sd_hyp\verb{[2]})}

\item{def_h_sd_hyp}{Vector of hyperparameters for def_h_sd ~ Cauchy(def_h_sd_hyp\verb{[1]}, def_h_sd_hyp\verb{[2]})}

\item{def_a_sd_hyp}{Vector of hyperparameters for def_a_sd ~ Cauchy(def_a_sd_hyp\verb{[1]}, def_a_sd_hyp\verb{[2]})}

\item{p_zero_home_hyp}{Vector of hyperparameters for p_zero_home ~ Beta(p_zero_home_hyp\verb{[1]}, p_zero_home_hyp\verb{[2]})}

\item{p_zero_away_hyp}{Vector of hyperparameters for p_zero_away ~ Beta(p_zero_away_hyp\verb{[1]}, p_zero_away_hyp\verb{[2]})}

\item{iter}{Number of iterations used by \code{rstan::sampling}}

\item{...}{Other arguments passed to \code{rstan::sampling}}

\item{chain}{Number of chains used by \code{rstan::sampling}}
}
\value{
An object of class \code{stanfit} returned by \code{rstan::sampling}
}
\description{
Bayesian regression for football championships data using a second negative binomial model
}
