% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footstan-package.R
\docType{package}
\name{footstan-package}
\alias{footstan-package}
\alias{footstan}
\title{The 'footstan' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (NA). RStan: the R interface to Stan. R package version 2.32.3. https://mc-stan.org
}
