% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_sample_data.R
\name{generate_sample_data}
\alias{generate_sample_data}
\title{Simulate n different championships}
\usage{
generate_sample_data(simulations, num_teams, model, params)
}
\arguments{
\item{simulations}{Number of simulations to generate}

\item{num_teams}{Number of teams in the championship.}

\item{model}{Which model to use to simulate the data}

\item{params}{Model parameters to simulate the data}
}
\value{
A list whose each element corresponds to a simulated championship
}
\description{
Simulate n different championships
}
