% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{gm}
\alias{gm}
\title{Fitting Goal Models with STAN}
\usage{
gm(data, family = "poisson", time = c("none", "dv"),
  homefield = c("intercept", "team"), date = lubridate::now(), ...)
}
\arguments{
\item{data}{A list or data frame containing the data. Should contain the
columns \code{HomeTeam}, \code{AwayTeam} giving the names of the teams
that play each game, together with their scores \code{HomeScore} and
\code{AwayScore}. If time dependence is used, the list must contain
\code{Time}, the time when each game was played.}

\item{family}{The goal model distribution. Defaults to "poisson", which is
the only distribution supported right now. Will probably implement
Negative Binomial in the future.}

\item{time}{Time dependence. Choose "none" for no time dependence or "dv" for
for decreasing variance time dependence with an inverse logit link.}

\item{homefield}{The kind of homefield advantage. Can be the same for all
teams or team specific.}

\item{date}{Optional argument telling the date of the match you wish to
predict. Only used when \code{time = "dv"}.}

\item{...}{Passed to \code{rstan::sampling}.}
}
\value{
A \code{stanfit} object.
}
\description{
\code{gm} is is used to fit goal models, specified by giving the desired
    family, time dependence, and a suitably formatted data frame.
}
\examples{
\dontrun{
# Ordinary Poisson goal model with Eliteserien after 2016.
gm(data = dplyr::filter(eliteserien, Season > 2016),
   family = "poisson")

# Decreasing variance Poisson goal model with Eliteserien after 2016.
gm(data = dplyr::filter(eliteserien, Season > 2016),
   family = "poisson",
   time = "dv",
   chains = 1)
}
}
