% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{goals_histogram}
\alias{goals_histogram}
\title{Makes histograms for the number of goals predicted by footstan::predict}
\usage{
goals_histogram(
  prediction_data,
  game,
  team = c("both", "home", "away"),
  home_team = "Home Team",
  away_team = "Away Team",
  fill_color = "blue",
  bin_width = 1
)
}
\arguments{
\item{prediction_data}{An object with the predictions made by the function footstan::predict}

\item{game}{Desired game in the new data}

\item{team}{Which team the user wants to visualize the estimated goals}

\item{fill_color}{Fill color for the histogram's bar.}

\item{bin_width}{Bin width for the histogram}

\item{home_name}{Name of the home team for visualization purposes.}

\item{away_name}{Name of the away team for visualization purposes.}
}
\value{
A ggplot2 plot
}
\description{
Makes histograms for the number of goals predicted by footstan::predict
}
