% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{plot_predictions}
\alias{plot_predictions}
\title{Makes histograms for the number of goals predicted by footstan::predict}
\usage{
plot_predictions(predictions, team_names = NULL)
}
\arguments{
\item{predictions}{An object with the predictions made by the function footstan::predict}

\item{team_names}{A data frame with ids and names to translate the ids into names to the plot}
}
\value{
A ggplot2 plot
}
\description{
Makes histograms for the number of goals predicted by footstan::predict
}
