% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict}
\alias{predict}
\title{Predicting new data using an already fitted model}
\usage{
predict(fit, new_data)
}
\arguments{
\item{fit}{A footstanfit object.}

\item{new_data}{List with the championship data.}
}
\value{
A vector with generated samples from the fitted model
}
\description{
Predicting new data using an already fitted model
}
