% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{prediction_proportions}
\alias{prediction_proportions}
\title{Predicting new data using an already fitted model}
\usage{
prediction_proportions(prediction_data)
}
\arguments{
\item{prediction_data}{An object with the predictions made by the function footstan::predict}
}
\value{
A data frame with the proportions of wins, draws and loses for the home team in the new_data
}
\description{
Predicting new data using an already fitted model
}
