% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{prediction_summary}
\alias{prediction_summary}
\title{Print summary of the predictions}
\usage{
prediction_summary(y1, y2)
}
\arguments{
\item{y1}{All the goals in home}

\item{y2}{All the goals away}
}
\description{
Print summary of the predictions
}
