% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_championship.R
\name{simulate_championship}
\alias{simulate_championship}
\title{Simulate a single championship}
\usage{
simulate_championship(num_teams, model, params)
}
\arguments{
\item{num_teams}{Number of teams in the championship.}

\item{model}{Which model to use to simulate the data}

\item{params}{Model parameters to simulate the data}
}
\value{
A data frame where each row it's a game, with the teams that are playing and the result of the game
}
\description{
Simulate a single championship
}
