% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{trace_plot}
\alias{trace_plot}
\title{Plot the draws for each chain}
\usage{
trace_plot(fit, param)
}
\arguments{
\item{fit}{A footstanfit object.}

\item{param}{Desired param to be graphed out}
}
\value{
A GGplot plot
}
\description{
Plot the draws for each chain
}
