% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footstan-package.R
\docType{package}
\name{footstan-package}
\alias{footstan-package}
\alias{footstan}
\title{Footstan: Fitting goal models in STAN.}
\description{
Programs to fit goal models in STAN.
}
\references{
Stan Development Team (2018). RStan: the R interface to Stan. R package version 2.18.1. http://mc-stan.org
}
