% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{visual_diagnostic}
\alias{visual_diagnostic}
\title{Plot diagnostic graphics for the MCMC chains}
\usage{
visual_diagnostic(fit, params)
}
\arguments{
\item{fit}{A footstanfit object.}

\item{params}{A character vector with the desired parameters to be ploted}
}
\value{
A GGplot plot
}
\description{
Plot diagnostic graphics for the MCMC chains
}
